Specify version types to be uint16 explicitly.

+++ domain/consensus/utils/txscript/example_test.go
@@ -37,7 +37,7 @@ func ExamplePayToAddrScript() {
 	}
 	fmt.Printf("Script Hex: %x\n", script.Script)
 
-	disasm, err := txscript.DisasmString(uint32(script.Version), script.Script)
+	disasm, err := txscript.DisasmString(script.Version, script.Script)
 	if err != nil {
 		fmt.Println(err)
 		return
